safari-534.53-branch/LayoutTests/canvas/philip/tests/2d.text.measure.width.empty.html

<!DOCTYPE html>
<title>Canvas test: 2d.text.measure.width.empty</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<style>
@font-face {
  font-family: CanvasTest;
  src: url("../fonts/CanvasTest.ttf");
}
</style>
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<span style="font-family: CanvasTest; position: absolute; visibility: hidden">A</span>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

deferTest();
setTimeout(wrapFunction(function () {
    ctx.font = '50px CanvasTest';
    _assertSame(ctx.measureText("").width, 0, "ctx.measureText(\"\").width", "0");
}), 500);


});
</script>
