LayoutTests/canvas/philip/tests/2d.text.measure.width.empty-expected.txt

< [index] >
2d.text.measure.width.empty
The empty string has zero width
References: 2d.text.measure
Actual output:
Passed
